Made topics and code corrections for better robot interaction

+++ ros_ws/src/navegation/src/KF.py
@@ -3,7 +3,7 @@
 #ros
 import rospy
 from geometry_msgs.msg import Pose2D
-from std_msgs.msg import Float64
+from std_msgs.msg import Float32
 
 #Puzzlebot
 import numpy as np
@@ -129,8 +129,8 @@ def main():
 
     #oddometry subscriber
     pose_sub  = rospy.Subscriber("/WOMBAT/navegation/odometry", Pose2D, rawPoseCallback, queue_size=10)
-    left_sub  = rospy.Subscriber(l_speed, Float64, leftCallback, queue_size = 1)
-    right_sub = rospy.Subscriber(r_speed, Float64, rightCallback, queue_size = 1)
+    left_sub  = rospy.Subscriber(l_speed, Float32, leftCallback, queue_size = 1)
+    right_sub = rospy.Subscriber(r_speed, Float32, rightCallback, queue_size = 1)
     
     #estimation publisher
     KF_pub = rospy.Publisher("/WOMBAT/navegation/pose", Pose2D, queue_size = 10)
